Move file-system interface to file-system/models
